reformatting; move ScriptExecutor
