+++ examples/hello-world/job_api/pt/requirements.txt
@@ -2,3 +2,4 @@ nvflare~=2.5.0rc
 torch
 torchvision
 pytorch_lightning
+tensorboard
